Added vue-starter in the scaffolding category. (#1440)
A Vue.js starter kit that lets you focus on more programming and less configuration. Its a very simple template based on vue-cli which adds in a lot of simplicity to webpack configuration and config management. Also, adds a lot of pre configured tools and libraries to help in the development of any app using it.